requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
openpyxl
python-multipart
httpx
bcrypt
python-jose[cryptography]
